Initial basic commit
I'm probably going to have to parameterize everything by the FloatFormat
Though I did consider methods such as only givign the FloatFormat when converting to bits or whatnot. But it is simply too useful to have it around all the time, and makes my life simpler than having to parameterize all operations by it.