Remove radix
It would be good to support other radices, most notably Decimal, but I'm not really working on that for the most part. There's complexities to the implementation of decimal that I'm not handling, and am not really equipped to handle at this moment in time.
As well, it makes the implementation of the bitformat more complex to prove properties about (removing radix let me get rid of some annoying cast logic, etc.).