Initial fixed impl of exponentBits
I was using log2 which is defined as the *floor*, but for bits we really want the ceil. In practice, this wasn't an issue- for all the formats I defined the result would be the same. However, it meant I could not prove that the biased exponent was not all one bits in general.
This proof is cleaner and also uglier, will fix