Introduce StdFloatFormat
The primary part of this is introducing a requirement that max_exp = 2^n - 1. This is needed to make the validFloat proof of toBits nicely implementable.
The core issue is that of junk values. A non-power of two (-1) max_exp results in bit representations that are unused (junk). Theoretically we could add an extra constraint on FloatBits that they are never created, but it adds extra pain to the whole process. Especially, as when in reality, all the formats that I know of use an emax of the form 2^n - 1.
I've packed the isStandardExpRange with it, though they are not necessarily always linked, but if there's a reason to split the class we can do that in the future.