Impl basic version of Rat.flog2
This is needed to make computable versions of a lot of the ulp, rounding functions, and so on. The 'idealized' versions would use Reals but those aren't computable!
Thankfully, for the calculations, we don't need every bit of the logarithm, just the floor for the exponent range around the number.

Unsure if there's a cleaner way to implement this.
We could have used a div version like how I believe Nat.log is implemented, but implementing it in terms of Nat.log made various transformations a lot simpler to show.